Migrate from IBM Cloud Functions to GCP Cloud Run. (#4)
ICF was deprecated a while ago, and the function would stop running later this year. Now, the bot uses equivalent technologies in GCP. It is a Go app that can be built in a Docker image. The image is used as the image for a Cloud Run job. The job is hooked up to a Cloud Scheduler trigger to run it every hour. There is now no need for managing the GCP service account credentials used to write data to BigQuery because everything is now running in GCP. The service account for the job to use as its identity is specified while deploying it, and that's all that needs to be done for GCP credentials.