Make dynamic wrt donation campaigns.
Code now starts from the home page to find the current donation campaign web page URL. Then, it fetches that page and parses the donations from it. It also now parses the goal instead of having that hard-coded. This change should make it follow each donation campaign for each stage of the appeals process.